YOUTUBE/chat_requirements.txt
weaviate-client==3.25.2
python-dotenv==1.0.0
streamlit==1.32.0
protobuf>=3.20,<5.0
tenacity>=8.1.0,<9.0 